requirements.txt
python-dotenv==1.0.1
aiogram==3.19.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
asyncpg==0.30.0
pydantic-settings
pytz==2024.2
numpy
alembic
